+++ quri-parts/packages/qulacs/quri_parts/qulacs/circuit/__init__.py
@@ -41,11 +41,11 @@ from quri_parts.circuit.gate_names import (
     is_two_qubit_gate_name,
     is_unitary_matrix_gate_name,
 )
-from quri_parts.qiskit.circuit.transpile import transpile_ecr_gates
 
 from ..utils import cast_to_list
 from .compiled_circuit import compile_circuit, compile_parametric_circuit
 from .qulacs_circuit_converter import circuit_from_qulacs
+from .transpile import transpile_ecr_gates
 
 _single_qubit_gate_qulacs: Mapping[
     SingleQubitGateNameType, Callable[[int], qulacs.QuantumGateBase]
